Perform TockyKmeansRF Analysis with Optional ROC Evaluation

This function performs the TockyKmeansRF analysis across multiple runs and optionally evaluates the ROC curves and calculates the AUC for each run. It can handle a single number of clusters or a vector of cluster numbers. The importance scores are accumulated across runs, and data is plotted based on these scores.

Usage

TockyRandomForestAnalysis(
  trainData,
  testData,
  num_cluster = 16,
  k = 100,
  percentile = 0.5,
  ROC = FALSE,
  verbose = TRUE,
  ctrl_group = NULL,
  expr_group = NULL,
  iter.max = 100,
  nstart = 1
)

Arguments

trainData

Training dataset as a TockyPrepData object.

testData

Test dataset as a TockyPrepData object.

num_cluster

A single integer specifying the number of clusters to evaluate.

k

Integer, number of repetitions for the analysis.

percentile

Numeric, percentile threshold for importance scores (between 0 and 1).

ROC

Logical, whether to perform ROC analysis.

verbose

Logical indicating whether to print progress messages and outputs. Default is TRUE.

ctrl_group

The name of the control group within `sampledef`.

expr_group

The name of the experimental group within `sampledef`.

iter.max

the maximum number of iterations allowed. To be passed to kmeans.

nstart

the number of random sets to be used in each clustering. To be passed to kmeans.

Value

A list containing the accumulated importance scores, optional ROC results, AUC values, and plots.

Examples

if (FALSE) { # \dontrun{
result <- TockyRandomForestAnalysis(trainData, testData, num_cluster = 16, k = 100, ROC = TRUE)
} # }